+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 * Unreleased
+    * Replace `writeHourMinute()` into `writeHourMinute12()` (12-hour format
+      with prepending space), and `writeHourMinute24()` (24-hour format with
+      prepending 0).
 * v0.1 (2021-08-10)
     * Initial release.
 * (2021-08-09)